tests: move util functions into separate file
